Refactor find_all_comps function into utils module and update CLI tests
